Ok, trying to make objects work, not going well
I'm manually making a table and returning that as the "button"
constructor, when I should be calling the "new" method, but I'm not sure
where that's supposed to be defined...I'll find it somewhere, I'm sure.

The Awesome lib looks at the class signals, in a very deep callstack somewhere
in luaobject.c (or maybe it was luaclass.c?). Point is they call the new
method correctly with the given arguments, whereas I'm not doing that
yet. BUT this is progress, and I might have to diverge with how the
Awesome lib does it...