Make all of the structs we pass to Lua repr(c)
Otherwise it's UB to reinterpret their bits. This is mostly prudent for
Object and the concrete implemntations for each object (e.g ButtonState)

Its probably not needed for Class, and might inhibit optimizations, but
I'm going to keep it in just in case because I'll take a performance hit
if it means I don't have weird UB down the road.